Modularize tools-api
