+++ packages/tools-api/src/modules/decoder-module/domain/encoding.ts
