+++ packages/tools-api/src/modules/decoder-module/domain/plugins/utils.ts
@@ -1,4 +1,4 @@
-import type { Address, Chain, TokenConfig } from '../../config/types'
+import type { Address, Chain, TokenConfig } from '../../../../config/types'
 import type { Value } from '../DecodedResult'
 
 export function tokenAmount(
