+++ packages/tools-api/src/modules/decoder-module/domain/plugins/whitebitBatchPlugin.ts
@@ -1,5 +1,5 @@
 import { toFunctionSelector } from 'viem'
-import type { Address, Chain, TokenConfig } from '../../config/types'
+import type { Address, Chain, TokenConfig } from '../../../../config/types'
 import type { DecodedCall, Value } from '../DecodedResult'
 import type { NestedCall } from './types'
 import { BinaryReader } from './utils'
