+++ packages/tools-api/src/services/ApplicationModule.ts
@@ -1,0 +1,6 @@
+import type express from 'express'
+
+export interface ApplicationModule {
+  routers?: express.Router[]
+  start?: () => Promise<void> | void
+}
