+++ packages/tools-api/src/types.ts
@@ -1,2 +1,2 @@
-export type * from './domain/DecodedResult'
-export type { ApiQuery } from './domain/ApiController'
+export type * from './modules/decoder-module/domain/DecodedResult'
+export type { ApiQuery } from './modules/decoder-module/domain/ApiController'
